Applying fixes based on reports from PVS Studio

+++ libs/statistics/include/hpx/statistics/histogram.hpp
@@ -91,8 +91,8 @@ namespace hpx { namespace util {
                     // determine bin positions (their lower bounds)
                     for (std::size_t i = 0; i < this->num_bins_ + 2; ++i)
                     {
-                        this->bin_positions_[i] =
-                            minimum_ + (i - 1.0) * bin_size_;
+                        this->bin_positions_[i] = minimum_ +
+                            (static_cast<double>(i) - 1.0) * bin_size_;
                     }
                 }
 #if defined(HPX_MSVC_WARNING_PRAGMA)
